+++ CHANGELOG.md
@@ -28,6 +28,9 @@ A brief description of the categories of changes:
 * Nothing yet
 
 ### Fixed
+* (rules) Signals are properly received when using {obj}`--bootstrap_impl=script`
+  (for non-zip builds).
+  ([#2043](https://github.com/bazelbuild/rules_python/issues/2043))
 * (rules) Fixes python builds when the `--build_python_zip` is set to `false` on Windows. See [#1840](https://github.com/bazelbuild/rules_python/issues/1840).
 * (pip) Fixed pypi parse_simpleapi_html function for feeds with package metadata
   containing ">" sign